Separated repetitive code to utils.py
